moved action to main package to utilize simconnector
