+++ crates/types/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod types {
-    pub mod action;
     pub mod bundle;
     pub mod category;
     pub mod connector_settings;
